requirements.txt
# Core dependencies
chromadb>=0.4.22
langchain>=0.1.0
sentence-transformers>=2.2.2
numpy>=1.24.0

# CLI dependencies
click>=8.1.7
rich>=13.7.0
pyyaml>=6.0.1

# MCP Server dependencies
# Note: MCP SDK will be installed separately as it may require specific version

# Utilities
python-dotenv>=1.0.0
pydantic>=2.5.0

# Development dependencies
pytest>=7.4.3
black>=23.12.0
mypy>=1.7.0
pytest-asyncio>=0.21.1
pytest-cov>=4.1.0